Update subgraph to include added collaterals post-deployment
