+++ subgraph/package.json
@@ -4,7 +4,8 @@
   "license": "MIT",
   "version": "1.0.0",
   "scripts": {
-    "codegen": "pnpm graph codegen"
+    "codegen": "pnpm graph codegen",
+    "build": "pnpm graph build"
   },
   "devDependencies": {
     "@graphprotocol/graph-cli": "^0.97.0"
