+++ subgraph/src/BoldToken.mapping.ts
@@ -2,59 +2,65 @@ import { Address, BigInt, DataSourceContext } from "@graphprotocol/graph-ts";
 import {
   CollateralRegistryAddressChanged as CollateralRegistryAddressChangedEvent,
 } from "../generated/BoldToken/BoldToken";
-import { BorrowerOperations as BorrowerOperationsContract } from "../generated/BoldToken/BorrowerOperations";
+// import { BorrowerOperations as BorrowerOperationsContract } from "../generated/BoldToken/BorrowerOperations";
 import { CollateralRegistry as CollateralRegistryContract } from "../generated/BoldToken/CollateralRegistry";
-import { TroveManager as TroveManagerContract } from "../generated/BoldToken/TroveManager";
-import { Collateral, CollateralAddresses } from "../generated/schema";
-import { TroveManager as TroveManagerTemplate, TroveNFT as TroveNFTTemplate } from "../generated/templates";
+// import { TroveManager as TroveManagerContract } from "../generated/BoldToken/TroveManager";
+// import { Collateral, CollateralAddresses } from "../generated/schema";
+import { CollateralRegistry as CollateralRegistryTemplate, TroveManager as TroveManagerTemplate, TroveNFT as TroveNFTTemplate } from "../generated/templates";
+import { addCollateral } from "./shared/collateral";
 
-function addCollateral(
-  collIndex: i32,
-  totalCollaterals: i32,
-  tokenAddress: Address,
-  troveManagerAddress: Address,
-): void {
-  let collId = collIndex.toString();
+// function addCollateral(
+//   collIndex: i32,
+//   totalCollaterals: i32,
+//   tokenAddress: Address,
+//   troveManagerAddress: Address,
+// ): void {
+//   let collId = collIndex.toString();
 
-  let collateral = new Collateral(collId);
-  collateral.collIndex = collIndex;
+//   let collateral = new Collateral(collId);
+//   collateral.collIndex = collIndex;
 
-  let troveManagerContract = TroveManagerContract.bind(troveManagerAddress);
+//   let troveManagerContract = TroveManagerContract.bind(troveManagerAddress);
 
-  let addresses = new CollateralAddresses(collId);
-  addresses.collateral = collId;
-  addresses.borrowerOperations = troveManagerContract.borrowerOperations();
-  addresses.sortedTroves = troveManagerContract.sortedTroves();
-  addresses.stabilityPool = troveManagerContract.stabilityPool();
-  addresses.token = tokenAddress;
-  addresses.troveManager = troveManagerAddress;
-  addresses.troveNft = troveManagerContract.troveNFT();
+//   let addresses = new CollateralAddresses(collId);
+//   addresses.collateral = collId;
+//   addresses.borrowerOperations = troveManagerContract.borrowerOperations();
+//   addresses.sortedTroves = troveManagerContract.sortedTroves();
+//   addresses.stabilityPool = troveManagerContract.stabilityPool();
+//   addresses.token = tokenAddress;
+//   addresses.troveManager = troveManagerAddress;
+//   addresses.troveNft = troveManagerContract.troveNFT();
 
-  collateral.minCollRatio = BorrowerOperationsContract.bind(
-    Address.fromBytes(addresses.borrowerOperations),
-  ).MCR();
+//   collateral.minCollRatio = BorrowerOperationsContract.bind(
+//     Address.fromBytes(addresses.borrowerOperations),
+//   ).MCR();
 
-  collateral.save();
-  addresses.save();
+//   collateral.save();
+//   addresses.save();
 
-  let context = new DataSourceContext();
-  context.setBytes("address:borrowerOperations", addresses.borrowerOperations);
-  context.setBytes("address:sortedTroves", addresses.sortedTroves);
-  context.setBytes("address:stabilityPool", addresses.stabilityPool);
-  context.setBytes("address:token", addresses.token);
-  context.setBytes("address:troveManager", addresses.troveManager);
-  context.setBytes("address:troveNft", addresses.troveNft);
-  context.setString("collId", collId);
-  context.setI32("collIndex", collIndex);
-  context.setI32("totalCollaterals", totalCollaterals);
+//   let context = new DataSourceContext();
+//   context.setBytes("address:borrowerOperations", addresses.borrowerOperations);
+//   context.setBytes("address:sortedTroves", addresses.sortedTroves);
+//   context.setBytes("address:stabilityPool", addresses.stabilityPool);
+//   context.setBytes("address:token", addresses.token);
+//   context.setBytes("address:troveManager", addresses.troveManager);
+//   context.setBytes("address:troveNft", addresses.troveNft);
+//   context.setString("collId", collId);
+//   context.setI32("collIndex", collIndex);
+//   context.setI32("totalCollaterals", totalCollaterals);
 
-  TroveManagerTemplate.createWithContext(troveManagerAddress, context);
-  TroveNFTTemplate.createWithContext(Address.fromBytes(addresses.troveNft), context);
-}
+//   TroveManagerTemplate.createWithContext(troveManagerAddress, context);
+//   TroveNFTTemplate.createWithContext(Address.fromBytes(addresses.troveNft), context);
+// }
 
 export function handleCollateralRegistryAddressChanged(event: CollateralRegistryAddressChangedEvent): void {
   let registry = CollateralRegistryContract.bind(event.params._newCollateralRegistryAddress);
   let totalCollaterals = registry.totalCollaterals().toI32();
+
+  // Create CollateralRegistry template
+  let context = new DataSourceContext();
+  context.setBytes("address:collateralRegistry", event.params._newCollateralRegistryAddress);
+  CollateralRegistryTemplate.createWithContext(event.params._newCollateralRegistryAddress, context);
 
   for (let index = 0; index < totalCollaterals; index++) {
     let tokenAddress = Address.fromBytes(registry.getToken(BigInt.fromI32(index)));
